+++ docs/.vuepress/components/LegacyWarning.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="warning custom-block">
         <p class="custom-block-title">WARNING</p>
-        <p>This is documentation for <u>legacy 2.x.x</u> versions. For the most current version <a :href="$withBase('/')">click here.</a></p>
+        <p>This is documentation for <u>legacy</u> versions. For the most current version <a :href="$withBase('/')">click here.</a></p>
     </div>
 </template>