+++ markwon-core/src/main/java/io/noties/markwon/core/CoreProps.java
@@ -19,6 +19,11 @@ public abstract class CoreProps {
 
     public static final Prop<Boolean> PARAGRAPH_IS_IN_TIGHT_LIST = Prop.of("paragraph-is-in-tight-list");
 
+    /**
+     * @since 4.1.1
+     */
+    public static final Prop<String> CODE_BLOCK_INFO = Prop.of("code-block-info");
+
     public enum ListItemType {
         BULLET,
         ORDERED