+++ markwon-image/src/main/java/io/noties/markwon/image/svg/SvgSupport.java
@@ -1,5 +1,9 @@
 package io.noties.markwon.image.svg;
 
+import android.util.Log;
+
+import androidx.annotation.NonNull;
+
 /**
  * @since 4.0.0
  */
@@ -13,7 +17,9 @@ public abstract class SvgSupport {
             com.caverock.androidsvg.SVG.class.getName();
             result = true;
         } catch (Throwable t) {
-            t.printStackTrace();
+            // @since 4.1.1 instead of printing full stacktrace of the exception,
+            // just print a warning to the console
+            Log.w("MarkwonImagesPlugin", missingMessage());
             result = false;
         }
         HAS_SVG = result;
@@ -21,6 +27,15 @@ public abstract class SvgSupport {
 
     public static boolean hasSvgSupport() {
         return HAS_SVG;
+    }
+
+    /**
+     * @since 4.1.1
+     */
+    @NonNull
+    static String missingMessage() {
+        return "`com.caverock:androidsvg:*` dependency is missing, " +
+                "please add to your project explicitly if you wish to use SVG media-decoder";
     }
 
     private SvgSupport() {